Rename MainPage to ShootingPage
